+++ src/index.d.ts
@@ -13,3 +13,5 @@ export { useAuthGuard } from "./wrapper/useAuthGuard";
 export { useFetchServer } from "./wrapper/useFetchServer";
 export { useDateTime } from "./wrapper/useDateTime";
 export { useCurrency } from "./wrapper/useCurrency";
+export { useGoogleLogin } from "./wrapper/useGoogleLogin";
+export { useFacebookLogin } from "./wrapper/useFacebookLogin";