+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5] - 2025-01-28
+
+### Added
+- **useGoogleLogin**: Wrapper untuk Google Sign-In dengan auto SDK injection dan popup-based authentication
+- **useFacebookLogin**: Wrapper untuk Facebook Login dengan auto SDK injection dan popup-based authentication
+
 ## [0.0.4] - 2025-01-20
 
 ### Added
@@ -71,6 +77,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Repository](https://github.com/ibnushahraa/vue3-utils)
 - [Issues](https://github.com/ibnushahraa/vue3-utils/issues)
 
-[Unreleased]: https://github.com/ibnushahraa/vue3-utils/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/ibnushahraa/vue3-utils/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/ibnushahraa/vue3-utils/compare/v0.0.4...v0.0.5
+[0.0.4]: https://github.com/ibnushahraa/vue3-utils/compare/v0.0.3...v0.0.4
+[0.0.3]: https://github.com/ibnushahraa/vue3-utils/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/ibnushahraa/vue3-utils/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/ibnushahraa/vue3-utils/releases/tag/v0.0.1